Getting header and general landing page skeleton
Add background image, some styling around the header text, and a few
notes in the HTML about what's going where. Also looking at the current
site and this one and deciding where to put some of the things that are
laying around a little, but that we could move around to the footer or
what have you.

Not really focusing on making things customizable yet, just trying to
get a feel for different blocks and content.