Add footer links and menu model for them
A basic link model for html or image links since that's what's currently
used on the site. a small amount of code for the footer that will
probably be abstracted away to its own partial soon enough.

Things to consider are probably: figuring out what css class to apply
for size of footer items based on how many they are. probably some
simple modular arithmetic to determine a good number to break on. For
example if there are 5 elements use 1-5, etc.